requirements.txt
passlib==1.7.4
pycryptodome==3.11.0
PyYAML==6.0
